changed from ints to longs for some safety

+++ IO.h
@@ -1,34 +1,34 @@
 #ifndef IO_H
 #define IO_H
 
-extern int current_state;
+extern long current_state;
 extern char** ids;
-extern int ids_count;
-extern int ids_max;
+extern long ids_count;
+extern long ids_max;
 extern long program_k;
 extern vector<string> initializedVariables;
 
 void yyerror(char const *);
-int getIdNumIfExists(char* id);
+long getIdNumIfExists(char* id);
 void convertStringToNumberAndPutInRegister(char* num);
 
 void readAndStore(char* id){
   if (doesStringContainIdSeparator(id)) { // is id of form a[b]
     char* bId = strrchr(id, '|') + 1;
-    int index = (int)(bId - id - 1);
-    int bAddr = getIdNumIfExists(bId);
+    long index = (long)(bId - id - 1);
+    long bAddr = getIdNumIfExists(bId);
 
     char* aId = (char*) malloc(sizeof(char)*index);
     memset(aId, '\0', index+1);
     strncpy(aId, id, index);
     char* a0Id = concat(aId, "0");
-    int a0IdAddr = getIdNumIfExists(a0Id);
+    long a0IdAddr = getIdNumIfExists(a0Id);
 
-    int a0IdAddrLen = snprintf( NULL, 0, "%d", a0IdAddr );
+    long a0IdAddrLen = snprintf( NULL, 0, "%ld", a0IdAddr );
     char* a0IdAddrString = (char*) malloc( a0IdAddrLen + 1 );
-    snprintf( a0IdAddrString, a0IdAddrLen + 1, "%d", a0IdAddr );
+    snprintf( a0IdAddrString, a0IdAddrLen + 1, "%ld", a0IdAddr );
 
-    int tmpMemAddr = getFreeMemoryAddress();
+    long tmpMemAddr = getFreeMemoryAddress();
 
     convertStringToNumberAndPutInRegister(a0IdAddrString); // put a0 adress in register
     STORE(tmpMemAddr);
@@ -41,7 +41,7 @@ void readAndStore(char* id){
     GET();
     STOREI(tmpMemAddr);
   } else {
-    for(int i=0; i < ids_count; i++){
+    for(long i=0; i < ids_count; i++){
       if(strcmp(id, ids[i]) == 0){
         GET();
         STORE(i);
