+++ assignment.h
@@ -2,12 +2,12 @@
 #define ASSIGNMENT_H
 
 extern long program_k;
-extern int getIdNumIfExists(char* id);
+extern long getIdNumIfExists(char* id);
 extern vector<string> initializedVariables;
 extern map<string, int> forLoopsVariables;
 
-int getIdNumIfExistsForAssignment(char* id){
-  for(int i=0; i < ids_count; i++){
+long getIdNumIfExistsForAssignment(char* id){
+  for(long i=0; i < ids_count; i++){
       if(strcmp(id, ids[i]) == 0){
          return i;
       }
@@ -18,7 +18,7 @@ int getIdNumIfExistsForAssignment(char* id){
 void assignToVariable(char* id) {
   if(bisonDebug) printf("assign to: '%s'\n", id);
   LOAD(ids_count);
-  int assignId = -1;
+  long assignId = -1;
   if (doesStringContainIdSeparator(id)) { // does id is of form a[b]
       char* bId = strrchr(id, '|') + 1;
       if(forLoopsVariables.find(bId) == forLoopsVariables.end()) {
@@ -27,20 +27,20 @@ void assignToVariable(char* id) {
           yyerror(UNINITIALIZED_VARIABLE);
         }
       }
-      int index = (int)(bId - id - 1);
-      int bAddr = getIdNumIfExists(bId);
+      long index = (long)(bId - id - 1);
+      long bAddr = getIdNumIfExists(bId);
 
       char* aId = (char*) malloc(sizeof(char)*index);
       memset(aId, '\0', index+1);
       strncpy(aId, id, index);
       char* a0Id = concat(aId, "0");
-      int a0IdAddr = getIdNumIfExists(a0Id);
+      long a0IdAddr = getIdNumIfExists(a0Id);
 
-      int a0IdAddrLen = snprintf( NULL, 0, "%d", a0IdAddr );
+      long a0IdAddrLen = snprintf( NULL, 0, "%ld", a0IdAddr );
       char* a0IdAddrString = (char*) malloc( a0IdAddrLen + 1 );
-      snprintf( a0IdAddrString, a0IdAddrLen + 1, "%d", a0IdAddr );
+      snprintf( a0IdAddrString, a0IdAddrLen + 1, "%ld", a0IdAddr );
 
-      int tmpMemAddr = getFreeMemoryAddress();
+      long tmpMemAddr = getFreeMemoryAddress();
 
       convertStringToNumberAndPutInRegister(a0IdAddrString); // put a0 adress in register
       STORE(tmpMemAddr);
